VMInstaller.sh - V1.06
- Added: Warning before overwritting config files
- Added: Possibility of refusing to use the config files
- Added: Required actions are now correctly color coded in yellow, hopefully
- Added: More informations about what the script is doing.
- Fixed: VirtualBox Guest Additions couldn't reinstall properly and
hanged as it was prompting the user a Yes or No question : as they were
already installed, and we weren't updating them, it asked if it should
continue or not. As the script is mainly automated, the user wasn't
seeing this question and thus, the script hanged.
- Changed: Spliting-up of VMInstaller.sh file to multiples .sh and .conf
files.
- Changed: Some text were changed to make them easier to understand
- Removed: Sleep times. Really, they were useless and slowed the script
down even more. We've got no time to waste. And if it goes to fast for
you, logs files exist.

This commit fixes #2 as we ask the user if we should overwrite the
config file.
Regarding #3, we still ask the user to either stop the script if the
VBox Guest Addtions weren't installed before, or continue if they were.